@@ -2,13 +2,14 @@ import java.util.regex.Matcher
 
 plugins {
     id 'application'
-    id 'org.openjfx.javafxplugin' version '0.0.13'
+    id 'org.openjfx.javafxplugin' version '0.0.14'
+    id 'com.diffplug.spotless' version '6.23.3'
 }
 
 group 'de.unijena.cheminf'
-version '1.0.6.0'
+version '1.1.1.0'
 
-sourceCompatibility = 1.17
+sourceCompatibility = 1.21
 
 //Creates javadoc and sources jars
 java {
