@@ -1,21 +1,26 @@
 /*
  * MORTAR - MOlecule fRagmenTAtion fRamework
- * Copyright (C) 2022  Felix Baensch, Jonas Schaub (felix.baensch@w-hs.de, jonas.schaub@uni-jena.de)
+ * Copyright (C) 2024  Felix Baensch, Jonas Schaub (felix.baensch@w-hs.de, jonas.schaub@uni-jena.de)
  *
  * Source code is available at <https://github.com/FelixBaensch/MORTAR>
  *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
  *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
  *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
  */
 
 package de.unijena.cheminf.mortar.preference;
@@ -24,8 +29,9 @@ import de.unijena.cheminf.mortar.model.fragmentation.algorithm.ErtlFunctionalGro
 import de.unijena.cheminf.mortar.model.fragmentation.algorithm.SugarRemovalUtilityFragmenter;
 import de.unijena.cheminf.mortar.model.settings.SettingsContainer;
 import de.unijena.cheminf.mortar.model.util.FileUtil;
-import org.junit.Assert;
-import org.junit.Test;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
 
 import java.awt.Color;
 import java.io.File;
@@ -73,20 +79,20 @@ public class PreferenceContainerTest {
         SingleNumberPreference tmpPreference4 = new SingleNumberPreference("Layout parameter xy", 2.0);
         SingleTermPreference tmpPreference5 = new SingleTermPreference("Welcoming message", "Welcome to MORTAR");
 
-        Assert.assertTrue(tmpContainer.add(tmpPreference1));
-        Assert.assertFalse(tmpContainer.add(tmpPreference1));
-        Assert.assertTrue(tmpContainer.contains(tmpPreference1));
-        Assert.assertTrue(tmpContainer.add(tmpPreference2));
-        Assert.assertTrue(tmpContainer.add(tmpPreference3));
-        Assert.assertTrue(tmpContainer.add(tmpPreference4));
-        Assert.assertTrue(tmpContainer.add(tmpPreference5));
-        Assert.assertTrue(tmpContainer.getSize() == 5);
+        Assertions.assertTrue(tmpContainer.add(tmpPreference1));
+        Assertions.assertFalse(tmpContainer.add(tmpPreference1));
+        Assertions.assertTrue(tmpContainer.contains(tmpPreference1));
+        Assertions.assertTrue(tmpContainer.add(tmpPreference2));
+        Assertions.assertTrue(tmpContainer.add(tmpPreference3));
+        Assertions.assertTrue(tmpContainer.add(tmpPreference4));
+        Assertions.assertTrue(tmpContainer.add(tmpPreference5));
+        Assertions.assertEquals(5, tmpContainer.getSize());
         IPreference[] tmpSortedNameAscending = tmpContainer.getPreferencesSortedNameAscending();
-        Assert.assertEquals(tmpPreference4, tmpSortedNameAscending[0]);
-        Assert.assertEquals(tmpPreference1, tmpSortedNameAscending[1]);
-        Assert.assertEquals(tmpPreference2, tmpSortedNameAscending[2]);
-        Assert.assertEquals(tmpPreference3, tmpSortedNameAscending[3]);
-        Assert.assertEquals(tmpPreference5, tmpSortedNameAscending[4]);
+        Assertions.assertEquals(tmpPreference4, tmpSortedNameAscending[0]);
+        Assertions.assertEquals(tmpPreference1, tmpSortedNameAscending[1]);
+        Assertions.assertEquals(tmpPreference2, tmpSortedNameAscending[2]);
+        Assertions.assertEquals(tmpPreference3, tmpSortedNameAscending[3]);
+        Assertions.assertEquals(tmpPreference5, tmpSortedNameAscending[4]);
 
         System.out.println(tmpContainer.getGUID());
         System.out.println(tmpContainer.getTimeStamp());
@@ -101,11 +107,11 @@ public class PreferenceContainerTest {
 
         File tmpPreferenceContainerFile = new File(tmpContainer.getContainerFilePathname());
         PreferenceContainer tmpReloadedContainer = new PreferenceContainer(tmpPreferenceContainerFile);
-        Assert.assertArrayEquals(tmpContainer.getPreferences(), tmpReloadedContainer.getPreferences());
-        Assert.assertEquals(tmpContainer.getGUID(), tmpReloadedContainer.getGUID());
-        Assert.assertEquals(tmpContainer.getTimeStamp(), tmpReloadedContainer.getTimeStamp());
-        Assert.assertEquals(tmpContainer.toString(), tmpReloadedContainer.toString());
-        Assert.assertEquals(tmpContainer, tmpReloadedContainer);
+        Assertions.assertArrayEquals(tmpContainer.getPreferences(), tmpReloadedContainer.getPreferences());
+        Assertions.assertEquals(tmpContainer.getGUID(), tmpReloadedContainer.getGUID());
+        Assertions.assertEquals(tmpContainer.getTimeStamp(), tmpReloadedContainer.getTimeStamp());
+        Assertions.assertEquals(tmpContainer.toString(), tmpReloadedContainer.toString());
+        Assertions.assertEquals(tmpContainer, tmpReloadedContainer);
         System.out.println();
     }
     //
